tailor: make 'output $DEBUG $str' work
Debug messages printed with 'output $DEBUG' so far were never actually
printed because the '-d' flag set the outputlevel to 5 while DEBUG was
defined as 10. This means the output function always evaluated the if
condition to false ([[ 10 -le 5 ]] = false) when using
'output $DEBUG'.

Fix this by increasing the outputlevel to 10 if '-d' was given.

Change-Id: Ifa08fa9ddcbcc490bb812e8b3e6888551b583897