undertaker: do not add unknown symbols from white-/blacklists into mergeblockconf formula
If symbols on the whitelist do not exist in the model, they prevent
the SAT solver from generating a valid solution, as unknown variables
default to 0.

For example, our shipped whitelist contains CONFIG_DRM_I915_KMS which
was removed from Linux in 2015 (upstream commit fd930478fb79). Trying
to generate configurations for kernels after this commit using our
whitelist would always have failed.

Fix this by checking if items from the whitelist and blacklist are
present in the model before adding them to the formula, and emit a
warning if they aren't.

Change-Id: I08a7ec5b8240d48e3b094ce3c6a1af5ea4d74048