tailor: enable early tracing with systemd init
Newer Ubuntu systems do not use upstart anymore, but come with systemd
as their init system. To launch undertaker-traceutil at startup, we
need a service description in the right format and allow undertaker-
tracecontrol-prepare-ubuntu to copy it to the correct location.

Additionally, this change uses the already present
scripts/init-bottom/ORDER file from the original initrd and prepends
start-ftrace to it instead of overwriting it with our own (possibly
outdated) copy.

Change-Id: Iec6c85ccba04844631eb82c90161fbc9e1956b61