undertaker/kconfig-dumps: rebuild dumpconf if necessary
The model conversion in undertaker/kconfig-dumps did not
properly declare dumpconf as a dependency for satyr/
undertaker-kconfigdump so make check would sometimes fail
due to a stale binary.

Fix this by explicitely making dumpconf a dependency
for the targets in question and using $(MAKE) instead
of make to get directory changes right.

Change-Id: I558667f1c20b25766a858bcc095457dd2b69bbdd