dumpconf: use correct length for item name
Kconfig provides a definition of SYMBOL_MAXLENGTH for the
maximum length for a symbol name. In fact, some symbols
in Linux are longer than the current maximum of 50
characters in dumpconf (e.g.,
VGACON_SOFT_SCROLLBACK_PERSISTENT_ENABLE_BY_DEFAULT,
introduced by commit 1a336c934623).

For such symbols the attribution of dependencies and
properties like default values and selects is broken as the
item name is cut off at 50 characters for these properties
(but not for the 'Item' definition as this uses sym->name).

Let's use the 'official' maximum length instead of the
arbitrary 50 character limit.

Change-Id: Iba33ad38c36fe93bdcd6b7bb16e331d4c73f4e79