dumpconf: include original symbol name in CVALUE strings
So far, our models add the prefix 'CVALUE_' for
non-boolean/tristate items (e.g., default values for string
or integer options). This logic leads to items like 'CVALUE_1'
being present in the formula. However, the model cannot
differentiate between different sources of the constant
value, which means that a required value of 1 for one option
might interfere with a negation in a completely unrelated
option. For example, we might end up with something like
'CONFIG_CVALUE_1 && !CONFIG_CVALUE_1', making the whole
formula unsolvable.

To work around this, this change also includes the name of
the surrounding symbol into the CVALUE_-prefixed string, so
a symbol named 'BAR' with a default value of 1 and a symbol
named 'FOO' with a dependency on a value different to 1
would lead to a formula of
'CONFIG_CVALUE_BAR_1 && !CONFIG_CVALUE_FOO_1' instead.

Change-Id: I9f52d85c6e291f95ad125044f2c05925b8d3dd16