undertaker-tailor: provide or derive KASLR offset
With KASLR (kernel address space layout randomization) we can
no longer simply pass the addresses extracted from the trace
to addr2line as the address in the image will not match the
address in the trace.

Instead we need to find the offset used by the randomization
to resolve the addresses to source code lines. There are a
couple of options here:

- record function names and offsets instead of addresses:
  the kernel source code already provides a script at
  scripts/faddr2line which can resolve a function name to a
  source code line. This would require changes to the
  traceutil state machine.
- map recorded addresses back to their location in the kernel
  image when processing the trace. This only requires changes
  and a little calculation in undertaker-tailor.

We chose the latter option, which allows to provide an offset
via the command line. You could manually find out the offset
by looking up a symbol on the target system using
/proc/kallsyms and locating the same symbol in the kernel
image on disk. In order to make life easier we also have a
simple heuristic which takes the smallest traced address and
applies a mask to it leaving only the bits which can be
randomized by KASLR (the alignment requirement for the kernel
is 2MiB so we can drop all lower bits). From this address we
subtract the (fixed) start address of the text segment to get
the offset which is used for the whole text segment.

Change-Id: Ibb6eb748bea1b74f28f13ec488e9f142371d482f